no checkbox to delete image

+++ views/station.py
@@ -244,10 +244,15 @@ class StationImageDeleteView(PermissionRequiredMixin, FormView):
         return context
 
     def form_valid(self, form):
-        self.img.delete()
+        if not 'cancel' in self.request.POST:
+            self.img.delete()
         return super(StationImageDeleteView, self).form_valid(form)
 
     def get_success_url(self):
+        if 'cancel' in self.request.POST:
+            return (reverse('bimblog:station_detail',
+                kwargs={'build_slug': self.build.slug,
+                'stat_slug': self.stat.slug}))
         return (reverse('bimblog:station_detail',
             kwargs={'build_slug': self.build.slug, 'stat_slug': self.stat.slug}) +
             f'?img_deleted={self.title}')
